Hanyizhou explore-send request

+++ bookcosmos/screens/Library.js
@@ -1,6 +1,6 @@
-import { StyleSheet, Text, View, FlatList } from "react-native";
+import { StyleSheet, Text, View, FlatList, Alert } from "react-native";
 import React, { useState, useEffect } from "react";
-import { collection, onSnapshot } from "firebase/firestore";
+import { collection, onSnapshot, query, where } from "firebase/firestore";
 import { database } from "../firebase-files/firebaseSetup";
 import {
   getAllDocs,
@@ -8,16 +8,29 @@ import {
 } from "../firebase-files/firestoreHelper";
 import Swipeable from "react-native-gesture-handler/Swipeable";
 import CustomButton from "../components/CustomButton";
+import { FontAwesome } from "@expo/vector-icons";
 
-export default function Library({ navigation }) {
+export default function Library({ navigation, userId, isMyLibrary }) {
   const [books, setBooks] = useState([]);
 
   useEffect(() => {
-    // Define the books collection reference
-    const booksCollection = collection(database, "books");
+    let booksQuery;
+    // Define the query to fetch books for a specific user
+    if (isMyLibrary) {
+      booksQuery = query(
+        collection(database, "books"),
+        where("owner", "==", userId)
+      );
+    } else {
+      booksQuery = query(
+        collection(database, "books"),
+        where("owner", "==", userId),
+        where("isBookInExchange", "==", false)
+      );
+    }
 
     // Subscribe to the query
-    const unsubscribe = onSnapshot(booksCollection, (snapshot) => {
+    const unsubscribe = onSnapshot(booksQuery, (snapshot) => {
       const fetchedBooks = [];
       snapshot.forEach((doc) => {
         fetchedBooks.push({ id: doc.id, ...doc.data() });
@@ -28,18 +41,45 @@ export default function Library({ navigation }) {
 
     // Clean up the subscription when the component unmounts
     return () => unsubscribe();
-  }, []);
+  }, [userId]);
 
   const handleDeleteItem = async (item) => {
     try {
-      // Call the deleteBookFromDB function to delete the book from the database
-      await deleteBookFromDB(item.id);
+      Alert.alert(
+        "Delete Book",
+        "Are you sure you want to delete this book?",
+        [
+          {
+            text: "Cancel",
+            style: "cancel",
+          },
+          {
+            text: "Delete",
+            onPress: async () => {
+              // Call the deleteBookFromDB function to delete the book from the database
+              await deleteBookFromDB(item.id);
 
-      // After successful deletion, fetch the updated list of books from the database
-      const updatedBooksData = await getAllDocs("books");
-      setBooks(updatedBooksData);
+              // After successful deletion, fetch the updated list of books from the database
+              const updatedBooksData = await getAllDocs("books");
+              setBooks(updatedBooksData);
+            },
+          },
+        ],
+        { cancelable: true }
+      );
     } catch (error) {
       console.error("Error deleting book:", error);
+    }
+  };
+
+  const handlePressBook = (item) => {
+    if (isMyLibrary) {
+      navigation.navigate("Add A Book", { editMode: true, bookId: item.id });
+    } else {
+      navigation.navigate("Book Detail", {
+        bookId: item.id,
+        ownerId: item.owner,
+      });
     }
   };
 
@@ -55,16 +95,12 @@ export default function Library({ navigation }) {
       )}
     >
       <View style={styles.item}>
-        <CustomButton
-          onPress={() =>
-            navigation.navigate("Add A Book", {
-              editMode: true,
-              bookId: item.id,
-            })
-          }
-        >
-          {item.bookName && <Text>Name: {item.bookName}</Text>}
-          {item.author && <Text>Author: {item.author}</Text>}
+        <CustomButton onPress={() => handlePressBook(item)}>
+          {item.bookName && <Text>{item.bookName}</Text>}
+          {item.author && <Text>{item.author}</Text>}
+          {item.isBookInExchange && (
+            <FontAwesome name="exchange" size={24} color="red" />
+          )}
         </CustomButton>
       </View>
     </Swipeable>
